Master merge and splits UGUI functionality from GUIRoot into PageSwitcher.cs and UGUIRoot.cs
