+++ Behaviours/GUI/UGUIRoot.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8f3cd62da8988334e9759dd80a8c9cf8
+timeCreated: 1461352436
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
