--- Behaviours/GUI/ZText.cs
@@ -1,21 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
-using System.Collections;
-using System.Collections.Generic;
-
-[ExecuteInEditMode]
-public class ZText : MonoBehaviour {
-
-	public Text text { get { return GetComponent<Text>(); } }
-
-
-	void Start() {
-		
-	}
-
-	
-	void Update() {
-		
-	}
-	
-}
